iCal/requirements.txt
fastapi
uvicorn
caldav
python-dotenv
requests
